Implement the redirect
The tricky cases are that some pages, like Amazon Music, don't support
Smile, and other pages, like a product details page, won't load on
the Smile domain if the user isn't logged in. For these cases, Amazon will
redirect you back to the normal domain, leading to an infinite loop if the
extension keeps trying to send the user to Smile.

For the Chrome and Firefox extension, the WebExtensions API provides a
requestId, which we can use to detect when we have already attempted a
redirect. The Safari App Extension framework has no such notion, so the
workaround is to set a "donateandsmile" query parameter when we try the
redirect. If we get a request with the same query parameter, we know
that Amazon has sent us back to the normal domain, and we should not
attempt another redirect.

We are counting on Amazon to preserve the query parameter through their
redirect. From my testing, that seems to be a safe assumption for now.

One alternative would be to maintain a blacklist of pages that don't
support Smile, but that would require manually keeping the list up to
date, and it also doesn't handle the pages where Smile is supported, but
the user needs to log in.