Change the extension's bundle id
This fixes an issue with uploading the build to App Store Connect. It
said the extension's bundle id was bad because it should start with the
application's bundle id and not contain more than one period after the
application's bundle id. The old bundle id seemed to fulfill these
requirements, but according to
https://stackoverflow.com/a/33735750/1481479, the bundle id for an
extension should be in the form of "com.my.app.XXX"

The offending bundle id was generated by Xcode, and Apple should fix the
extension template.